virt: fix DIMM removal
In the case of DIMM removal, it is not saved in VDSM. Libvirt removes
the device, but since we don't save it, it will result in a warning
message and without updating the engine.

This patch will save the DIMM device and on device removal event will
check it. It will trigger the update to the engine and won't show a
false warning. It will take care to update the engine in cases of device
removal that didn't execute by VDSM or when VDSM restarted in the
middle of the action.

Change-Id: I0edb985a0805289a4680ed6997b79e93675ebecc
Bug-Url: https://bugzilla.redhat.com/1883483
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>