qga: mark guest-info check from single location
Instead of storing the information in every place where we call
guest-info command it is better to do it only once from the
_qga_capability_check() method. The patch changes the original behavior
slightly -- in _on_boot() function the information is now stored even in
case of failure. This is in fact correct. The old behavior was wrong and
did not match the rest of the poller.

Bug-Url: https://bugzilla.redhat.com/1834233
Change-Id: Ib4c93fd1680336ce72cd0d69b462dc1838766024
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>