qga: invalidate qemu-ga capabilities on reboot
This lets us pick up new capabilities after the reboot. In most cases
the capabilities will be the same, but this way we can handle unusual
situations. For example when the agent is updated shortly before reboot
or VMs with multi-boot.

It will also allow us to detect the UP state properly in the
following patch.

Bug-Url: https://bugzilla.redhat.com/1834233
Change-Id: If06f647dce21bea19b2f9b6ec0c5ba01fb220255
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>