backup.py: add backup_mode property for each backup disk
Add 'backup_mode' = 'full'/'incremental' for each disk that
participates in a VM backup.
This information will guide VDSM on the type of backup that should
be taken for each disk.

This new property will give the option to create a checkpoint that
contains both 'full' and 'incremental' backup disks.

So now the request data from Engine to VDSM contains the following
structure:
{
    backup_id: "backup-1",
    disks: [
        {
            "img_id": "disk1",,
            "domain_id": "domain1",
            "volume_id": "volume1",
            "checkpoint": false,
            "backup_mode": "full",
        },
        {
            "img_id": "disk2",,
            "domain_id": "domain2",
            "volume_id": "volume2",
            "checkpoint": true,
            "backup_mode": "incremental",

        },
    from_checkpoint_id: "from_checkpoint_id",
    to_checkpoint_id: "to_checkpoint_id"
}

The 'backup_mode' property will be used in the followup patch.

Change-Id: I7d60bb006e71f76a33b12baca081315523aee0a5
Bug-Url: https://bugzilla.redhat.com/1861674
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>