backup.py: Use 'backup_mode' property for each disk in the backup
There are two supported backup modes:

  - 'full':
        should be used for 'raw' disks and for
        newly added disks that weren't part of a
        previous checkpoint.

  - 'incremental':
        should be used for disks that were
        part of a previous backup, this mode
        can be used only when taking an
        incremental backup.

This patch uses the newly reported backup_mode that the engine provides.
In order to have backward compatibility, if the backup mode doesn't
reported by the engine, the backup XML will not include it and the
backup will not be able to contain both 'full' and 'incremental' backups
under the same checkpoint.

Change-Id: I23a65c63a062ecb642eec40c7136d1afba7c278a
Bug-Url: https://bugzilla.redhat.com/1861674
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>