net, ovs: Enforce MAC of base interface
When vdsm takes over existing interface it should ensure
the MAC address is cloned to the OvS interface. This way
the IP address obtained from DHCP server will remain
the same as it was configured on the base interface.

This was causing issues mostly for host deploy flows.
Engine was not able to communicate with host because
the ip address has changed.

Only exception that is not covered is when we are creating
bond and network on top of that bond in the same request.
This scenario is complicated by the nature of some bond
modes and it is hard to reliably determine which MAC address
will be assigned to the newly created bond.

Change-Id: I77a1dee7b28b05e6d9da432b0e14cbf274d75865
Bug-Url: https://bugzilla.redhat.com/1809102
Signed-off-by: Ales Musil <amusil@redhat.com>