virt: Fix migration destination detection in recovery
When VM status recovery detects a paused VM in migration, it assumes
that it's an outgoing migration.  But it can be also an incoming
migration, which is actually much more likely, because a VM is paused
only for a short time (unless it's a post-copy migration, which is
handled at a different place) on the source while it's paused for long
time on the destination.

Let's distinguish between source and destination by examining the
migration job, which is present only on the source.  The implemented
solution is still racy and cause trouble, but only under the following
circumstances (all the conditions must happen):

- Vdsm gets restarted on the migration source during a migration.
- The recovery is invoked during the short time the VM is paused for
  the final stage of a pre-copy migration.
- The migration fails.
- The migration job disappears between checking the VM status and the
  immediately following job check.

This scenario is very unlikely and can be remedied by Vdsm restart if
it ever happens, so it's probably not worth of complicating the code.

Change-Id: Iea3eaa0854c18612a2a6aa551107814ebcbd0780
Bug-Url: https://bugzilla.redhat.com/1877632
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>