tool: libvirt: Stop libvirt sockets on reconfiguration
When reconfiguring libvirtd with vdsm-tool, we stop the
'libvirtd.service'. This however is not enough - since libvirt
switched to socket activation, even when the main 'libvirtd.service'
daemon is not running, one can respawn it by making a connection
to one of its sockets.

This patch adds libvirt's sockets to the list of units that are being
stopped and started during reconfiguration so we can avoid this issue.

Change-Id: Ia55eab45095e5ace58cdc89a13f5418250942c5d
Bug-Url: https://bugzilla.redhat.com/1878724
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>