storage.hsm: rename whiteList variable and function
The purpose of "whiteList" in __cleanStorageRepository() is not apparent
by its name. In fact, it is a list of glob patterns of files we should
keep as they are. This patch renames the whiteList to KEEP_PATTERNS, to
make it more apparent that it is actually a constant listing globs of
files we want to keep on disk. Similarly, the patch renames the function
that checks if a file is covered by the globs and thus should be kept.

__cleanStorageRepository() begs for a more thorough refactoring. E.g, I
see no reason to modify "dirs" while iterating over it. However, this
patch intentionally limits itself to name changes only.

Change-Id: Ie3293f989d2c58deeb2a1f2527361ca2aad8f0bd
Signed-off-by: Dan Kenigsberg <danken@redhat.com>