bridge_test: rename apiWhitelist
apiWhitelist is badly named. It is in camel case; it does not show that
it is a local variable; it does not show that it is constant; it claims
to be a list, while it is actually a tuple; it is not clear what is its
purpose.

This patch renames it to _COPIED_API_OBJECTS to fix the above issues.

Change-Id: I5bdc95eceb84c935bea3fcf3fb7536c9250532dd
Signed-off-by: Dan Kenigsberg <danken@redhat.com>