qga: consume guest-get-disks command
We can now get the list of all disks with new guest-get-disks command.
This includes also disks without mounted filesystem. Since this is a new
feature we don't want to rely solely on this new command as it is likely
it won't be supported by most of the guests in the foreseeable future.
We try to fetch info first using this command, if this is successful
this information is kept. Otherwise we resort to the previous way and
gather the information from data of guest-get-fsinfo.

Bug-Url: https://bugzilla.redhat.com/1836661
Change-Id: I3aec1cbd9d0e5237d3611dca7f2c9787e00dc976
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>