volume.py: introduce add/remove bitmap for a volume
New SDM.add_bitmap will be used to start an offline cold backup
for a volume that is part of a powered-off VM.

The new add_bitmap API will add a bitmap to the given volume,
this dirty bitmap will be exposed later via qemu-nbd server.

Request for SDM.add_bitmap
{
  'job_id': job-uuid,
  'vol_info':
    {
      'sd_id': sd-uuid,
      'img_id': image-uuid,
      'vol_id': volume-uuid,
      'generation': 1
    }
  'bitmap': 'bitmap1'
}

The new remove_bitmap API will remove a bitmap from the given volume.
This is currently not used by the engine but may be needed in
'vdsm-client' tool.

Request for SDM.remove_bitmap -
{
  'job_id': job-uuid,
  'vol_info':
    {
      'sd_id': sd-uuid,
      'img_id': image-uuid,
      'vol_id': volume-uuid,
      'generation': 1
    }
  'bitmap': 'bitmap1'
}

Change-Id: Ib406b7bb132044f7f446d1cf55597c0671098ac9
Bug-Url: https://bugzilla.redhat.com/1891470
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>