tests: Fix check for qemu-img bitmap support
qemu-img bitmap is available in qemu-img 4.2.0-29 on CentOS 8.2, but
--merge works only in RHEL 8.3.0 nightly (qemu img 5.1.0). Since we use
bitmaps operations only in in cluster version 4.5, requiring libvirt
6.6 (requires qemu-img 5.1), it is simpler to allow bitmaps operation
only with qemu-img 5.1.

To make our life more interesting, --merge does not work in qemu-img
5.1.0 on Fedora, breaking the tests when running locally.

Change the bitmaps_supported() check to use qemu-img version, and add a
new @requires_bitmaps_merge_support marker to skip merge tests on
Fedora. Tests for merging bitmaps use the new marker.

We will be able to eliminate these checks when CentOS 8.3 will be
released, and qemu-img 5.2 will be available in Fedora.

Change-Id: I9aed2485416fa6e1bee5f1c4d1c5684139507a05
Signed-off-by: Nir Soffer <nsoffer@redhat.com>