tests: Mark failing copy_data tests
Seems that recent changes in qemu-img after 4.2.0-19 cause the target
image to be sparse when using -n option.

Here is an example flow:

$ qemu-img create -f raw -o preallocation=falloc src.raw 1m
Formatting 'src.raw', fmt=raw size=1048576 preallocation=falloc

$ qemu-img create -f raw -o preallocation=falloc dst.raw 1m
Formatting 'dst.raw', fmt=raw size=1048576 preallocation=falloc

$ qemu-io -f raw -c 'write -P 240 0 65536' src.raw
wrote 65536/65536 bytes at offset 0
64 KiB, 1 ops; 00.06 sec (1005.353 KiB/sec and 15.7086 ops/sec)

$ ls -lhs
total 2.0M
1.0M -rw-r--r--. 1 nsoffer nsoffer 1.0M Oct 25 19:49 dst.raw
1.0M -rw-r--r--. 1 nsoffer nsoffer 1.0M Oct 25 19:49 src.raw

$ qemu-img convert -t none -T none -f raw -O raw -n src.raw dst.raw

$ ls -lhs
total 1.1M
 64K -rw-r--r--. 1 nsoffer nsoffer 1.0M Oct 25 19:46 dst.raw
1.0M -rw-r--r--. 1 nsoffer nsoffer 1.0M Oct 25 19:46 src.raw

This is a real issue in the system we need to fix, but we need to fix
the build first. Add a new marker for this issue, and mark the affected
tests as expected failures.

Change-Id: Ief2df47af4cbfafe1267a92b19479c434f51990d
Bug-Url: https://bugzilla.redhat.com/1891520
Signed-off-by: Nir Soffer <nsoffer@redhat.com>