nbd: Fix compatibility with older engines
When backing_chain support was added in:

commit 372ffa7f0b6343fa86fe17163f9062d4a5edbe67
Author: Nir Soffer <nsoffer@redhat.com>
Date:   Wed Sep 16 12:37:58 2020 +0300

    nbd: Allow exporting a volume without the backing chain

backing_chain property was added with default=True:

    backing_chain = properties.Boolean(default=True)

This works fine if the backing_chain property is not initialized, but
when the caller does not specify the value, it was initialized to None:

    self.backing_chain = config.get("backing_chain")

Setting property to None overrides the default value, so the property
became None, which is treated as False later, leading to wrong NBD
export returning only the specified volume instead of the entire chain.

This breaks older engine that do not specify the backing_chain argument.

Fixed by treating unspecified value as True. This may be fixed in the
properties module, but this requires much more work and thinking, and we
need to fix this requestion quickly.

Change-Id: Icffcace76960ed3b202199515447f4c5dd23d61a
Bug-Url: https://bugzilla.redhat.com/1892403
Signed-off-by: Nir Soffer <nsoffer@redhat.com>