nbd: Support dirty bitmaps
Add bitmap argument to NBDServerConfig. When set, export the specified
bitmap in qemu-nbd. NBD client can use the dirty bitmap to download
changed blocks when a VM is powered off.

The current code works for the simple case of a single volume, and can
be used for initial testing with engine.

Issues:

- Support volume chain: create an overlay and merge the bitmaps from all
  volumes to the overlay.

- Handle errors like requesting bitmaps for raw volume, missing bitmaps
  or invalid bitmap.

- Using private nbd module from ovirt_imageio. This module should become
  public in future imageio release.

Change-Id: I0427a75bcd6514efee518ad2d5fd345f28c8a358
Signed-off-by: Nir Soffer <nsoffer@redhat.com>