qemuimg: Report backing file format
This is useful for tests and for code that want to inspect the backing
chain. The value is returned only if the image was created with the -F
argument. Creating images with backing files without specifying the
backing format is deprecated and logs a warning. We need to clean up
some test code using this shortcut.

The new key "backingformat" is similar format to the "backingfile" key
for consistency.

Change-Id: Ibe0b11e72b2a0ffa827a7ff73cca2c745255bd6e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>