tool: Handle template parent volumes in dump-volume-chains
Previously we used StoargeDomain.getVolumes() which included also
templates volumes in the response. Now we build the volumes list
per image in the client, and we did not consider the case of
template volumes.

This is fixed by a two pass processing over the volumes result
from the Storage domain dump API; on first pass we iterate
the volume info results and group them into volumes_info dict
by the image id of each volume. On second pass we iterate the
resulted volume info volumes by images and add the parent volume
info from the volumes result if it is missing in the volumes_info
for that particular image, indicating it has a parent template
volume from another image.

Output before fix:

   image:    a3749e70-1598-49d4-86a0-2d6ccb78a106

             Error: no volume with a parent volume Id _BLANK_UUID found e.g: (a<-b), (b<-c)

             Unordered volumes and children:

             - 9fcaac13-652a-4d23-a5ff-7162ab3a87b7 <- ffa30cde-7346-437b-923c-2cfd4e3d6216
               status: OK, voltype: LEAF, format: COW, legality: LEGAL, type: SPARSE, capacity: 107374182400, truesize: 1073741824

Output after fix:

   image:    a3749e70-1598-49d4-86a0-2d6ccb78a106

             - 9fcaac13-652a-4d23-a5ff-7162ab3a87b7
               status: OK, voltype: SHARED, format: COW, legality: LEGAL, type: SPARSE, capacity: 107374182400, truesize: 3758096384

             - ffa30cde-7346-437b-923c-2cfd4e3d6216
               status: OK, voltype: LEAF, format: COW, legality: LEGAL, type: SPARSE, capacity: 107374182400, truesize: 1073741824

Base volume is shared from another image:

    "9fcaac13-652a-4d23-a5ff-7162ab3a87b7": {
        "apparentsize": 3758096384,
        "capacity": 107374182400,
        "ctime": 1596547254,
        "description": "",
        "disktype": "DATA",
        "format": "COW",
        "generation": 0,
        "image": "45563a94-6498-4c42-ba00-8930a51c81ca",
        "legality": "LEGAL",
        "mdslot": 3,
        "parent": "00000000-0000-0000-0000-000000000000",
        "status": "OK",
        "truesize": 3758096384,
        "type": "SPARSE",
        "voltype": "SHARED"
    }

Change-Id: I27b5394b53b042ed6480f8330fed3cdf00a90a3f
Bug-Url: https://bugzilla.redhat.com/1839444
Signed-off-by: Amit Bawer <abawer@redhat.com>