virt: fix memory reporting when NVDIMM is attached
Memory of the VM is reported incorrectly when NVDIMM is attached to the
VM. This stems from the fact that <memory> element in libvirt XML
includes not only physical memory but also size of all NVDIMMs. To make
the situation more complicated, the initial XML delivered from Engine
does not include NVDIMMs but only physical memory.

While it is not possible to migrate VM with NVDIMM at the moment the
patch tries to be future-proof in this aspect. If the VM is migrated
then we assume that the memory in source XML already includes the NVDIMM
sizes.

Bug-Url: https://bugzilla.redhat.com/1893773
Change-Id: I495d43f5359df2a8346034a2c302f41e958dc440
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>