backup.py: add require_consistency flag to BackupConfig
Adding the new require_consistency flag will let the backup app
to choose if the backup will fail if the VM failed to freeze or not.

If the VM failed to freeze it might lead to inconsistent data on the
backup so in order to prevent it require_consistency should be set to
'True'.

The default value for require_consistency flag is 'False'.

Change-Id: I9d8469793af88f14815c31ab80d711f86a63ff0c
Bug-Url: https://bugzilla.redhat.com/1894413
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>