virt: add xml source argument and property to domain descriptor
The `initial=True` covered the case of initial VM start as well as
migration where XML source is libvirt on source host. Instead of just
having a one flag it is better to separate these two cases and thus keep
more information about the source of the XML. That way we can
distinguish if the initial XML comes from Engine or from migration
source. DomainDescriptor does not really care about this distinction,
but there is a follow-up patch where we need to handle only VM start and
not migrations.

Bug-Url: https://bugzilla.redhat.com/1893773
Change-Id: Ife02aff5087a5a725b7a0b747cddd47937b99063
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>