qemuimg: Return qemu-img info output as is
We used to report only some of the info returned by qemu-img info,
change the names, and move some format specific data to the top level of
the dict.

In general, modifying the output of the tool in a tool wrapper is a bad
idea. Instead of reusing the knowledge of the real tool when working
with the wrapper, now we need to work with 2 different formats. This
makes no sense.

Moving format-specific data to top level is useful, saving some typing
in code accessing these keys, but it does not worth the effort of
maintaining the wrapper and modifying it each time a new key is needed.

Now we return the actual output from qemu-img info as is.

Change-Id: I25384d3a17a88aacc853232259da2ecb78b125e1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>