tests: fix metadata_storage_test
Running metadata_storage_test with python 3.8, many tests fail with

    E   AssertionError: '<?xm[53 chars]vice devtype="disk" name="vda">\n        <devi[689 chars]m>\n' != '<?xm[53 chars]vice name="vda" devtype="disk">\n        <devi[689 chars]m>\n'
    E     <?xml version='1.0' encoding='utf-8'?>
    E     <vm>
    E   -     <device devtype="disk" name="vda">
    E   +     <device name="vda" devtype="disk">

as the identifying attrs is a dict with keys "devtype" and "name" (in
this order), in this cases specified in
_get_drive_conf_identifying_attrs(), while all expected XML samples use
ordering "name" and "devtype". Fix the expected XML samples to use
correct ordering.

Tests are passing on CI as it used older python version. Starting python
3.8, XML attributes are kept in the same order as provided by user,
while in previous versions attributes were orders lexically, see [1] for
more details.

[1] https://bugs.python.org/issue34160

Change-Id: I69e8741b9e0fbba43502b0f1f02ad685781e6369
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>