vm: remember UP state after power up
When the VM starts and we notice that the QEMU Guest Agent is up we
consider the VM in UP state. This state is however not remembered in the
VM object which holds a separate state status. If the agent disappears
or fails for some reason shortly after the boot the UP state is not
remembered and VM goes back into POWERING_UP until the time out is
reached and the VM is considered "definitely UP".

This behavior is bad and once we proclaim the VM as UP it should stay up
and not fall back to POWERING_UP. This is solved by remembering the
state in the VM object. Ideally we should not keep two or three
different pieces of information about the VM state scattered throughout
the code, but that calls for a larger refactoring.

The issue is easily triggered when backup is invoked immediately when
the VM gets to UP state. This calls fsfreeze in the guest which disables
some of the commands in the agent (like guest-get-users). When we try to
call such commands it results in a failure and we don't consider the
agent in good state for some time. During this window the state from
agent is not considered valid which results in the fallback to VM
internal state (which can still be POWERING_UP). While our broken
handling of agent after fsfreeze is normally not an issue (when done
later during VM lifetime) it should also be fixed, but that is for a
separate patch too.

Bug-Url: https://bugzilla.redhat.com/1893656
Change-Id: I0c8f80ca36e22378844e441c06465c565638adf2
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>