xml: provide function for ordering XML attributes
Prior python 3.8, XML attributes were re-oreded in lexical order when
parsed. This has changed in python 3.8 (see [1]) and attributes are
kept in the same order as provided by user.

Example in Python 3.6.8:

    Python 3.6.8 (default, Apr 16 2020, 01:36:27)
    [GCC 8.3.1 20191121 (Red Hat 8.3.1-5)] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> import xml.etree.ElementTree as ET
    >>> ET.tostring(ET.fromstring("<test arg2='2' arg1='1' />"))
    b'<test arg1="1" arg2="2" />'

Same in Python 3.8.6:

    Python 3.8.6 (default, Sep 25 2020, 00:00:00)
    [GCC 10.2.1 20200723 (Red Hat 10.2.1-1)] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> import xml.etree.ElementTree as ET
    >>> ET.tostring(ET.fromstring("<test arg2='2' arg1='1' />"))
    b'<test arg2="2" arg1="1" />'

Add a function which provides ordering of attributes and thus can mimic
old behaviour. This is useful mainly for the tests which rely on
automatic ordering of attributes and now fail with python 3.8. So we
either have to order attributes in lexical order or adjust all XML
samples which we use for test asserts.

As shown above, importing as re-exporting XML via ElementTree doesn't
work and minidom doesn't work as well:

    Python 3.8.6 (default, Sep 25 2020, 00:00:00)
    [GCC 10.2.1 20200723 (Red Hat 10.2.1-1)] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> from xml.dom.minidom import parseString
    >>> parseString("<test arg2='2' arg1='1' />").toxml()
    '<?xml version="1.0" ?><test arg2="2" arg1="1"/>'

Unless I missed some functionality/option which enables backward
compatible lexical ordering, custom function for attribute ordering
seems to be the only way how to do it.

[1] https://bugs.python.org/issue34160

Change-Id: I05ca49b058060241efd6658bdd400170c1fdffd0
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>