virt: Support for reading and writing TPM data
TPM is much more useful with TPM data persistence than without it.
TPM emulation is handled by swtpm.  swtpm stores its data under
libvirt directory for each running VM.  It uses data already present
when the VM starts or new data is created if it's not present.  The
data is deleted when a VM domain is undefined.  This patch introduces
basic functions for reading and writing the TPM data for a given VM.

There are important security considerations.  Supervdsm must be used
to access TPM data because access to it is restricted.  Any user who
can access supervdsmd or vdsmd can read and write the data, which is
inherently needed to be able to transfer the data to or from Engine.
We restrict supervdsm access to swtpm subdirectories, with properly
formatted UUID names, to prevent changes outside it by calling Vdsm
APIs.  We must also be cautious not to leak parts of TPM data when
logging exceptions.  Although untrusted persons shouldn't have any
access to hosts, there can still be QEMU bugs etc. and it's good to
have basic restrictions.

More TPM data processing functionality will be added, especially
interaction with Engine, in followup patches.

Change-Id: I79826723c7abeafae2d7888198bc81921debddeb
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>