backup.py: set stop_backup as a success if VM is not defined
If VM.stop_backup called for an undefined VM (can happen if the VM
powered-off for the guest for example), stopping the backup will fail
and the scratch disks that were created for that backup will not be
removed.

This patch handles this by detecting the case when the VM is done cleanup
the scratch disks and handle the operation as a success.

Change-Id: I5ffdfe37d0b40e1a6ffb6aab61548e5a5a44dd4b
Bug-Url: https://bugzilla.redhat.com/1900518
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>