virt: Write TPM data obtained from Engine
When a VM with a TPM device is started, Engine should provide TPM
data.  If the data is not provided, typically for new VMs or when a
TPM device is added to the VM for the first time, initial data is
created automatically by swtpm.

This patch adds support for a new VM.create parameter that contains
TPM data to start the VM with.  Support for reading TPM data and
reporting it back to Engine will be added in followup patches.

Change-Id: I3b48316ae93ac78d71a71c9638406023554696af
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>