virt: Monitor TPM data
This patch introduces periodic reading TPM data.  Data is considered
stable if its reported hash doesn't change in two last consecutive
calls.  Only in such a case, or if forced reading is requested,
typically after the VM is down, the data reported to Engine is
updated.  This is to protect against data changes performed at the
same time the data is read.

We compute a cryptographic hash of the data for Engine, not to leak
any information about TPM data in logs etc.

Reporting the read TPM data to Engine will be addressed in followup
patches.

Change-Id: Iadd7db5c5404457e2e33e1e88e3d437063e701f8
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>