virt: New verb VM.getExternalData
It serves to transfer TPM data to Engine.  In future, it may serve to
transfer other similar data, such as secure boot data.

The verb must be flexible enough to serve not only for transferring
different kinds of data (TPM, secure boot) but also to transfer it
under different circumstances and reliability/efficiency requirements.
TPM data is retrieved and sent unconditionally after each request by
default.  This is to support the first use case: Sending the final
data when the VM is down, before it is (together with its data)
undefined by libvirt.  Future Engine patches may utilize more relaxed
approaches allowing periodic monitoring and retrieving and
transferring the data only when needed.

It is important to handle errors properly.  Otherwise Engine may fail
to retrieve the data or it may fail to destroy the VM.

Change-Id: I0e527c04f5a5c1eed71a01f87e2af0e0f5e692bf
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>