common: Allow protection of fields other than "password"
Not every sensitive piece of information can be reasonably named
"password".  For instance, TPM data, secure boot data, or their hashes
are all pieces of information that shouldn't be logged.

Include into the password protection all items with keys starting with
"_X_" so that any piece of data can be protected from logging.

Change-Id: Ie4fb846944a200eb650351c16b54f29c93dff38d
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>