vmdevices: handle mapping of luns with image id
LUNs were not set with serial. This caused to libvirt auto-set it as the
disk alias. We changed it into the LUN ID, but this harms backwards
compatibility. Therefore, we now aligned with non-LUN disks, having the
serial set with the image ID. The only different case is LUN with
pass-through, where we can't set the serial and it will exist as the LUN
ID. As part of this change, the mapping of LUNs need to handle the new
serial set by the engine, in order to provide the logical name of that
device.

Change-Id: Ic1541585297832a86fd64302e148c96468f41f0d
Bug-Url: https://bugzilla.redhat.com/1904774
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>