backup.py: allow redefine checkpoints using backup configuration
Libvirt doesn't require now the VM domain XML when checkpoint redefines.

VDSM can compose the checkpoint XML given the configuration
of the backup that was taken for that checkpoint.

This patch allows to redefine a checkpoint using one of the two ways:
  1. By providing the checkpoint XML (as string)
  2. By composing the checkpoint XML on the host using the backup
     configuration.

If both of the fields are given, the checkpoint will redefine using the
backup configuration. If none of the fields are given the operation will
be failed.

The support for redefining a checkpoint using the checkpoint XML remains
in order to support old engine that does not support sending backup
configuration.

Change-Id: Iaf6e4bcc97cefdef2cdead922bec0f52b9b12c4e
Bug-Url: https://bugzilla.redhat.com/1901835
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>