virt: vm: wrap external data into a class
The methods for handling TPM can be reused for other external data
(NVRAM coming soon). To make adding new external data easier this patch
wraps everything into an object and collects the external data in a
dictionary.

Rather than trying to handle all data kinds together we prefer to focus
on one kind at a time. That way the flows can be kept separate and in
case returning some data fails or blocks for some reason it won't
interfere with retrieval of other kinds.

Bug-Url: https://bugzilla.redhat.com/1669178
Change-Id: I04e7ce8ed19953b628bb1d9b7408cf38f32ca688
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>