net, ovs: Add MTU support for OvS networks
MTU on OvS is due to it's nature enforced on the
the network itself and it's base interface. The OvS
bridge adjusts it's MTU based on ports. If there
are multiple networks over single interface with
different MTU values, the MTU applied to the base
interface is highest of all networks.

Change-Id: I425178a69e887624e49c097e5ddd7ee44323347d
Signed-off-by: Ales Musil <amusil@redhat.com>