virt: supervdsm: add support for reading and writing NVRAM data
Add functionality to read and write NVRAM data. Files are stored in
libvirt directory. Normally the directory is created by libvirt. In rare
cases the directory may not exist yet and we have to create it. This is
probably needed only when the first UEFI VM started on the host is the
one with Secure boot enabled and stored NVRAM.

Bug-Url: https://bugzilla.redhat.com/1669178
Change-Id: I9a3d04cd03a0c38a5c1aebb5920a48a13f7e6ae0
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>