virt: sysprep: clean host name
virt-sysprep does not remove host name or set it to any generic value.
It only removes it from ifcfg network-scripts (which likely does nothing
at all). To avoid new VMs re-using the original host name we need to use
'--hostname <new-hostname>' to change it. But since sealing is done when
the template is created we don't know new host name yet. The only sane
option seams to be to change the host name to some generic value.

Change-Id: Ia068f7f7d8ab9ce01ae747c003fdd840c65de192
Bug-Url: https://bugzilla.redhat.com/1860492
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>