net, tests: Skip sysfs bonding tests on travis
The recent failure in Travis CI uncovered interesting thing.

Container is not able to load kernel module unless it is
privileged to do so. So any loading of bonding module directly
fails. But kernel is able to load the module on its own when we
request creation of new bond however the loaded module
has some flaws when done this way. It allows to create
bond via sysfs or iproute2 but fails with permission
error on modifying the bond options via sysfs.

Change-Id: I562c6a743a01e712c933e24839624e2f1d590702
Signed-off-by: Ales Musil <amusil@redhat.com>