constants.py: add backup scratch disk type to content type
When a backup is taken for a VM, for each disk that
participate in the backup a scratch disk is created.

This patch adds this disk type to the content type
so it will be easy to distinguish between the scratch
that was created during a backup and regular data disks.

Change-Id: Ia3f995946a5ca971d7d73a0e334061107e347466
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>