backup.py: allow using scratch disks on shared storage
Until now, when a backup created for a VM a scratch disk was
created for each disk that participates in the backup.
The scratch disks were created on the local storage of the host
that runs the VM.

The host local storage is limited and if the storage runs out during
the backup due to intensive data written to the VM, the VM will
be paused.

To avoid this situation, the scratch disks can be created on the
same shared storage where the backed-up disk resides. To allow that,
the engine needs to create the scratch disk and prepare it before
starting the backup, and teardown the disk and remove it when the backup
ended.

A New 'scratch_disk' attribute was added for each disk configuration
parameter in the backup, if will contain all the needed info on the scratch
disks so it can be used in the backup (disk type, path to the prepared disk).

For example, to start a backup for a VM with scratch disks that already
were created by the engine the following request structure should be used:

{
    backup_id: "backup-1",
    disks: [
        {
            "img_id": "disk1",,
            "domain_id": "domain1",
            "volume_id": "volume1",
            "checkpoint": false,
            "backup_mode": "full",
            "scratch_disk": {
                "path": "/path/to/disk1/scratch_disk"
                "type": "file",
             }
        },
        {
            "img_id": "disk2",,
            "domain_id": "domain2",
            "volume_id": "volume2",
            "checkpoint": true,
            "backup_mode": "incremental",
            "scratch_disk": {
                "path": "/path/to/disk2/scratch_disk"
                "type": "block",
            }
        },
    from_checkpoint_id: "from_checkpoint_id",
    to_checkpoint_id: "to_checkpoint_id"
}

Change-Id: Ie3860763ea26872e1cb95250e177bd7e74a32557
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>