nbd: Support exporting bitmap from backing chain
When exporting a bitmap from backing chain, we cannot use qemu-nbd
--bitmap option, since it exports only the bitmap from the top image.

To export bitmap from the entire chain, we create an overlay on top of
the actual volume, add empty bitmap to the overlay, and merge the bitmap
from all the backing chain nodes into the overlay bitmap. Finally we
export the overlay instead of the actual volume.

To allow this mode, path verification was changed to allow a transient
disk with backing file in the storage repository.

The operation may fail with new storage bitmap exceptions if the bitmap
is invalid or missing in some of the nodes, or does not exist in any
node. Engine should delete the bitmap from stored checkpoints when
handling these errors.

We already have bitmap exceptions in the generic vdsm exceptions module,
but they cannot be used in storage code, so I added new bitmap
exceptions in the storage exceptions module.

Change-Id: Ied8e53afce9f305b21e3a37bd1ce50dd6af7a206
Signed-off-by: Nir Soffer <nsoffer@redhat.com>